+++ device-bootloaders/src/ServiceProcessorBL2/driver/noc_configuration.c
@@ -34,6 +34,106 @@
 #include "hal_noc_reconfig.h"
 #include "slam_engine.h"
 #include "noc_configuration.h"
+#include "hwinc/noc_esr.h"
+#include "noc_reconfigure.h"
+
+/* Globals for NOC Shire Remap */
+static int g_displace = SPARE_SHIRE_BIT_POSITION;
+
+static int getPhysicalID(int virtualID)
+{
+    // Validate virtualID to avoid array index out of bounds
+    if (virtualID < 0 || virtualID > NUM_SHIRES - 1)
+    {
+        return -1;
+    }
+    else
+    {
+        return phy_id[virtualID];
+    }
+}
+
+static void swap_regs(uint32_t adbase[][NUM_SHIRES], unsigned int bridge_id, int displace,
+                      int spare, long unsigned int offset)
+{
+    int displace_phy_id = getPhysicalID(displace);
+    int spare_phy_id = getPhysicalID(spare);
+
+    if (displace_phy_id != -1 && spare_phy_id != -1)
+    {
+        const uint64_t baseAddr = R_SP_MAIN_NOC_REGBUS_BASEADDR;
+        uint64_t displace_base = baseAddr + adbase[bridge_id][displace_phy_id] + offset;
+        uint64_t displace_mask = baseAddr + adbase[bridge_id][displace_phy_id] + offset + 8;
+        uint64_t spare_base = baseAddr + adbase[bridge_id][spare_phy_id] + offset;
+        uint64_t spare_mask = baseAddr + adbase[bridge_id][spare_phy_id] + offset + 8;
+
+        // Swap Base values
+        uint64_t spare_base_org_value = *(uint64_t *)spare_base;
+        uint64_t displace_base_org_value = *(uint64_t *)displace_base;
+        *(uint64_t *)spare_base = displace_base_org_value;
+        *(uint64_t *)displace_base = spare_base_org_value;
+
+        // Swap Mask values
+        uint64_t spare_mask_org_value = *(uint64_t *)spare_mask;
+        uint64_t displace_mask_org_value = *(uint64_t *)displace_mask;
+        *(uint64_t *)spare_mask = displace_mask_org_value;
+        *(uint64_t *)displace_mask = spare_mask_org_value;
+    }
+    else
+    {
+        // Handle the case when virtual IDs are not found
+        Log_Write(LOG_LEVEL_ERROR, "Error: Virtual ID not found.\n");
+    }
+}
+
+static void swap_shires(int displace, int spare)
+{
+    for (unsigned int bridge_id = 0; bridge_id < NUM_NOC_ESR_BRIDGE; bridge_id++)
+    {
+        for (unsigned int min_shire = 0; min_shire < NOC_ESR_BRIDGE_ARRAY_COUNT; min_shire++)
+        {
+            swap_regs(adbase_noc_esr_bridge, bridge_id, displace, spare,
+                      min_shire * NOC_ESR_BRIDGE_ARRAY_ELEMENT_SIZE);
+        }
+    }
+
+    for (unsigned int bridge_id = 0; bridge_id < NUM_NOC_ESR_SIB_TOL3; bridge_id++)
+    {
+        for (unsigned int min_shire = 0; min_shire < NOC_ESR_SIB_TOL3_ARRAY_COUNT; min_shire++)
+        {
+            swap_regs(adbase_noc_esr_sib_tol3, bridge_id, displace, spare,
+                      min_shire * NOC_ESR_SIB_TOL3_ARRAY_ELEMENT_SIZE);
+        }
+    }
+
+    for (unsigned int bridge_id = 0; bridge_id < NUM_NOC_ESR_SIB_TOSYS; bridge_id++)
+    {
+        for (unsigned int min_shire = 0; min_shire < NOC_ESR_SIB_TOSYS_ARRAY_COUNT; min_shire++)
+        {
+            swap_regs(adbase_noc_esr_sib_tosys, bridge_id, displace, spare,
+                      min_shire * NOC_ESR_SIB_TOSYS_ARRAY_ELEMENT_SIZE);
+        }
+    }
+
+    for (unsigned int bridge_id = 0; bridge_id < NUM_NOC_ESR_BRIDGE_IOS; bridge_id++)
+    {
+        swap_regs(adbase_noc_esr_bridge_ios, bridge_id, displace, spare, 0);
+    }
+
+    for (unsigned int bridge_id = 0; bridge_id < NUM_NOC_ESR_BRIDGE_PS; bridge_id++)
+    {
+        swap_regs(adbase_noc_esr_bridge_ps, bridge_id, displace, spare, 0);
+    }
+
+    for (unsigned int bridge_id = 0; bridge_id < NUM_NOC_ESR_BRIDGE_MEM; bridge_id++)
+    {
+        for (unsigned int mem_shire = 0; mem_shire < NOC_ESR_BRIDGE_MEM_ARRAY_COUNT; mem_shire++)
+        {
+            swap_regs(adbase_noc_esr_bridge_mem, bridge_id, displace, spare,
+                      mem_shire * NOC_ESR_BRIDGE_MEM_ARRAY_ELEMENT_SIZE);
+        }
+    }
+}
 
 static void reconfig_memory_shire(CSR_SLAM_TABLE *reconfig_table_ptr)
 {
@@ -129,6 +229,59 @@ int32_t NOC_Configure(uint8_t mode)
         reconfig_noc_1_minshires
     */
     reconfig_minion_shire(CSR_SLAM_TABLE_PTR_NULL);
+
+    return SUCCESS;
+}
+
+int32_t Set_Displace_Shire_Id(uint64_t shire_mask)
+{
+    if (shire_mask == SHIRE_MASK_DEFAULT)
+    {
+        // All is good, no remap required
+        Log_Write(LOG_LEVEL_INFO, "Using the default shire mask, no shire to be displaced.\n");
+        return SUCCESS;
+    }
+
+    // Unless we're using the default mask, spare shire bit must be set and
+    // there can be only one disabled shire.
+    if ((HIGHEST_SET_BIT_POSITION(shire_mask) != SPARE_SHIRE_BIT_POSITION) ||
+        (NUM_ENABLED_SHIRES(shire_mask) != (NUM_SHIRES - 1)))
+    {
+        Log_Write(LOG_LEVEL_ERROR, "Error: Invalid shire mask\n");
+        return ERROR_INVALID_ARGUMENT;
+    }
+
+    g_displace = DISPLACE_SHIRE(shire_mask);
+
+    if ((g_displace < 0) || (g_displace > MAX_SHIRE_BIT_POSITION))
+    {
+        return ERROR_INVALID_ARGUMENT;
+    }
+
+    return SUCCESS;
+}
+
+int32_t Get_Displace_Shire_Id(void)
+{
+    return g_displace;
+}
+
+int32_t Get_Spare_Shire_Id(void)
+{
+    return SPARE_SHIRE_BIT_POSITION;
+}
+
+int32_t NOC_Remap_Shire_Id(int displace, int spare)
+{
+    if (displace == spare)
+    {
+        // All is good, no remap required
+        Log_Write(LOG_LEVEL_INFO, "Using the default shire mask, no remap needed.\n");
+        return SUCCESS;
+    }
+
+    Log_Write(LOG_LEVEL_INFO, "Remapping shire %d.\n", displace);
+    swap_shires(displace, spare);
 
     return SUCCESS;
 }