+++ device-bootloaders/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [SW-19082] Maxion configuration functions.
 - Added support to read GPT tables
+- [SW-18887] Added NOC shire re-map functionality
 ### Changed
 ### Deprecated
 ### Removed